rename configs dir to conf
