+++ src/conf/settings.py
@@ -21,7 +21,7 @@ AUTH_METHOD = "PASSWORD"
 AUTH_METHODS = [ "PASSWORD" ]
 
 # The user's username and password will be saved in the file below.
-CONFIG_FILE = "configs/configs.ini"
+CONFIG_FILE = "conf/configs.ini"
 DEFAULT_CONFIG_FILE = CONFIG_FILE
 
 PASSWORD = "PASSWORD"
