+++ src/core/TapipyController.py
@@ -1,6 +1,6 @@
 import shared.options.handlers
 
-from configs import settings
+from conf import settings
 from core.Controller import Controller
 from core.Authenticator import Authenticator as Auth
